Fix NULL pDstWaitStageMask
If we have a non-zero amount of semaphores to wait on, when submitting
to queue, we also need a list of wait stages masks, one for each
semaphore.

In order to avoid allocating a list for every frame however, I've opted
for a static list of masks with a maximum of `MAX_PRESENT_SWAP_SEMAPHORE_COUNT`.
I've guessed 16 being enough, though this can be changed, and guessing that a
stage of `VK_PIPELINE_STAGE_TRANSFER_BIT` is the stage we want it to wait
on.

That does mean however, if we have more than 16, then our list won't be
long enough. To avoid Vulkan from trying to read outside of valid
memory, then in this case, I've opted to just revert to the old
behaviour.